Proper Rasterizer/requirements.txt
PyOpenGL==3.1.6
pygame==2.1.2
numpy==1.23.4
PyYAML==6.0
perlin-noise==1.12